Vlaue management

+++ ExplorerFM/FieldAttribute.cs
@@ -6,31 +6,31 @@ namespace ExplorerFM
     {
         public string Name { get; }
         public bool IsTripleIdentifier { get; }
-        public long Min { get; }
-        public long Max { get; }
+        public int Min { get; }
+        public int Max { get; }
         public bool IsSql { get; }
 
         public FieldAttribute(string name)
-            : this(name, false, long.MinValue, long.MaxValue, true)
+            : this(name, false, int.MinValue, int.MaxValue, true)
         { }
 
         public FieldAttribute(string name, bool isTripleIdentifier)
-            : this(name, isTripleIdentifier, isTripleIdentifier ? 0 : default(long), long.MaxValue, true)
+            : this(name, isTripleIdentifier, isTripleIdentifier ? 0 : default(int), int.MaxValue, true)
         { }
 
-        public FieldAttribute(string name, long min)
-            : this(name, false, min, long.MaxValue, true)
+        public FieldAttribute(string name, int min)
+            : this(name, false, min, int.MaxValue, true)
         { }
 
-        public FieldAttribute(string name, long min, long max)
+        public FieldAttribute(string name, int min, int max)
             : this(name, false, min, max, true)
         { }
 
-        public FieldAttribute(string name, long min, long max, bool isSql)
+        public FieldAttribute(string name, int min, int max, bool isSql)
             : this(name, false, min, max, isSql)
         { }
 
-        private FieldAttribute(string name, bool isTripleIdentifier, long min, long max, bool isSql)
+        private FieldAttribute(string name, bool isTripleIdentifier, int min, int max, bool isSql)
         {
             Name = name;
             IsTripleIdentifier = isTripleIdentifier;
